// src/app/maps/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
    type:        string;
    query:       string[];
    features:    Feature[];
    attribution: string;
}

export interface Feature {
    id:            string;
    type:          string;
    place_type:    string[];
    relevance:     number;
    properties:    Properties;
    text_es:       string;
    language_es?:  Language;
    place_name_es: string;
    text:          string;
    language?:     Language;
    place_name:    string;
    bbox:          number[];
    center:        number[];
    geometry:      Geometry;
    context:       Context[];
}

export interface Context {
    id:          ID;
    short_code:  ShortCode;
    wikidata:    Wikidata;
    text_es:     Text;
    language_es: Language;
    text:        Text;
    language:    Language;
}

export enum ID {
    Country8754 = "country.8754",
    Region115762 = "region.115762",
    Region17458 = "region.17458",
}

export enum Language {
    Es = "es",
}

export enum ShortCode {
    Co = "co",
    CoAnt = "CO-ANT",
    CoCal = "CO-CAL",
}

export enum Text {
    Antioquia = "Antioquia",
    Caldas = "Caldas",
    Colombia = "Colombia",
}

export enum Wikidata {
    Q123304 = "Q123304",
    Q230607 = "Q230607",
    Q739 = "Q739",
}

export interface Geometry {
    type:        string;
    coordinates: number[];
}

export interface Properties {
    wikidata?: string;
}